requirements.txt
bitly-api==0.3
boto3==1.7.65
botocore==1.10.65
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.0.11
docutils==0.14
Flask==1.0.2
idna==2.5
imageio==2.1.2
InstagramAPI==1.0.2
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
moviepy==0.2.3.2
newsapi==0.1.1
newsapi-python==0.2.3
ngrok==0.0.1
nltk==3.3
numpy==1.15.0
Pillow==5.2.0
PyJWT==1.6.4
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
requests==2.11.1
requests-toolbelt==0.7.0
s3transfer==0.1.13
six==1.11.0
tqdm==4.11.2
twilio==6.15.1
urllib3==1.21.1
Werkzeug==0.14.1
